+++ base/sysimg.jl
@@ -166,7 +166,8 @@ push!(I18n.CALLBACKS, Help.clear_cache)
 
 # sparse matrices and linear algebra
 include("sparse.jl")
-include("jlsparse.jl")
+importall .SparseMatrix
+include("matrixmarket.jl")
 include("linalg.jl")
 importall .LinAlg
 include("broadcast.jl")
